+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Add a component title & description to the top of the configuration dialog
 * Make description & title show up in the `fv-list` element
 * Sort nodes in the `fv-list` alphabetically
+* Fix a schema clone issue
 
 ## 0.3.1 (2022-06-12)